+++ shared_albums/request_models.go
@@ -1,0 +1,5 @@
+package shared_albums
+
+type shareTokenBody struct {
+	ShareToken string `json:"shareToken"`
+}