+++ shared_albums/request_options.go
@@ -1,0 +1,6 @@
+package shared_albums
+
+type ListOptions struct {
+	PageSize                 int  `url:"pageSize"`
+	ExcludeNonAppCreatedData bool `url:"excludeNonAppCreatedData"`
+}